requirements_enhanced.txt
# Enhanced AI Ticker Analyzer - Additional Requirements for RAG and LangGraph

# Base requirements from requirements.txt should be installed first

# RAG and Vector Store dependencies
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.21.0

# LangGraph and workflow orchestration
langgraph>=0.0.40
langchain>=0.1.0
langchain-core>=0.1.0

# Visualization and plotting
matplotlib>=3.6.0
networkx>=3.0
plotly>=5.15.0

# Enhanced text processing (optional but recommended)
transformers>=4.21.0

# Async support enhancements
asyncio-mqtt>=0.13.0  # For potential async extensions

# Development and testing (optional)
pytest-asyncio>=0.21.0

# Note: Install with: pip install -r requirements_enhanced.txt